refactor(pvmc): clean the compilation flow (#122)
* feat(pvm): introduce IR in parser

* feat(pvm): introduce instruction info

* feat(pvm): introduce ir block reader

* feat(pvm): introduce IR verifier

* chore(clippy): make clippy happy

* fix(pvm): duplicated insertion of functions

* feat(pvm): collect undiscovered functions

* fix(pvm): two-pass for IR parsing

* feat(pvm): resolve all jumps

* feat(pvmc): batch exports from entry in ir

* chore(pvm): disable analysis in pvmi

* chore(pvm): support parsing one big function

* chore(pvmc): get back the tests of pvmc

* chore(pvm): get back stack to compiler

* chore(pvmc): get back jump table

* chore(pvm): use registers to pass values for branching

* feat(pvmc): get back variables for registers

* chore(pvmc): remove outdated tests

* chore(pvmc): host the trap block

* feat(pvmc): introduce macro blocks

* chore(pvmc): update the arguments of sbrk

* chore(pvm): logs for debugging gas

* refactor(pvmc): unified gas meter

* chore(pvmc): introduce log api

* refactor(pvmc): remove dynamic dispatching

* feat(pvm): get back the compilation cache

* chore(pvmc): remove stack usages

* refactor(pvmc): remove clif cache

* feat(pvm): rename jastime to spacevm

* chore(clippy): make clippy happy

* chore(vm): remove outdated tests

+++ crates/pvm/interpreter/src/visitor.rs
@@ -15,7 +15,6 @@ impl Visitor for Interpreter {
     fn visit_add_32(&mut self, format: format::RRR, _range: &Range<usize>) -> Result<()> {
         let format::RRR { reg0, reg1, reg2 } = format;
         let value = (self.rget(reg0) as u32).wrapping_add(self.rget(reg1) as u32) as u64;
-
         self.rset(reg2, value.sign_ext32());
         Ok(())
     }
@@ -371,7 +370,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_u8(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: u8 = self.read((addr + imm0) as u32)?;
+        let value: u8 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value as u64);
         Ok(())
     }
@@ -379,7 +378,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_u16(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: u16 = self.read((addr + imm0) as u32)?;
+        let value: u16 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value as u64);
         Ok(())
     }
@@ -387,7 +386,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_i16(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: i16 = self.read((addr + imm0) as u32)?;
+        let value: i16 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value as u64);
         Ok(())
     }
@@ -395,7 +394,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_u32(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: u32 = self.read((addr + imm0) as u32)?;
+        let value: u32 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value as u64);
         Ok(())
     }
@@ -403,7 +402,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_i32(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: i32 = self.read((addr + imm0) as u32)?;
+        let value: i32 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value.as_u64());
         Ok(())
     }
@@ -411,7 +410,7 @@ impl Visitor for Interpreter {
     fn visit_load_ind_u64(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let addr = self.rget(reg1);
-        let value: u64 = self.read((addr + imm0) as u32)?;
+        let value: u64 = self.read((addr as i64 + imm0 as i64) as u32)?;
         self.rset(reg0, value);
         Ok(())
     }
@@ -1042,31 +1041,37 @@ impl Visitor for Interpreter {
     ) -> Result<()> {
         let format::RII { reg0, imm0, imm1 } = format;
         let address = self.rget(reg0);
-        self.write((address + imm0) as u32, imm1)
+        self.write((address as i64 + imm0 as i64) as u32, imm1)
     }
 
     fn visit_store_ind_u8(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let address = self.rget(reg1);
-        self.write((address + imm0) as u32, self.rget(reg0) as u8)
+        self.write((address as i64 + imm0 as i64) as u32, self.rget(reg0) as u8)
     }
 
     fn visit_store_ind_u16(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let address = self.rget(reg1);
-        self.write((address + imm0) as u32, self.rget(reg0) as u16)
+        self.write(
+            (address as i64 + imm0 as i64) as u32,
+            self.rget(reg0) as u16,
+        )
     }
 
     fn visit_store_ind_u32(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let address = self.rget(reg1);
-        self.write((address + imm0) as u32, self.rget(reg0) as u32)
+        self.write(
+            (address as i64 + imm0 as i64) as u32,
+            self.rget(reg0) as u32,
+        )
     }
 
     fn visit_store_ind_u64(&mut self, format: format::RRI, _range: &Range<usize>) -> Result<()> {
         let format::RRI { reg0, reg1, imm0 } = format;
         let address = self.rget(reg1);
-        self.write((address + imm0) as u32, self.rget(reg0))
+        self.write((address as i64 + imm0 as i64) as u32, self.rget(reg0))
     }
 
     fn visit_sub_32(&mut self, format: format::RRR, _range: &Range<usize>) -> Result<()> {
